+++ tools/compilers/gcc-glibc/glibc-2.14.1-ros/sysdeps/ros/recvfrom.c
@@ -5,7 +5,7 @@
 #include <sys/socket.h>
 #include <ros/syscall.h>
 
-int __recvfrom (int s, void *buf, size_t len, int flags,
+ssize_t __recvfrom (int s, void *buf, size_t len, int flags,
                  struct sockaddr *from, socklen_t *fromlen)
 {
 	return ros_syscall(SYS_recvfrom, s, buf, len, flags, from, fromlen);
